pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agente"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "flask",
    "pandas",
    "openpyxl",
    "rapidfuzz",
    "numpy",
]
